requirements.txt
ayla-iot-unofficial-crevetor==1.3.2
aiohttp>=3.9
